Remove large blank space at the bottom on mobile. Fixes #104.
When using skrollr on mobile and at least one skrollable was not positioned
absolut or fixed (for example when using relative mode, where elements stay in
document flow), there was a space at the bottom of the page. The height of the
space was exactly the height of the viewport.
The cause was the wrong order when initializing iScroll. It was initialized
BEFORE we set the correct style on the body-element (position absolute, full
width and height). Thus when setting the styles on the body and iScroll had
already been instantiated the new height of the body (100%) was added to the
content height causing the blank space.