Properly get the scrollHeight cross-browser. Fixes #131 and #133
We were using body.scrollHeight which does not always hold the correct value.
We are now using the Math.max of five different values to properly get the
scrollHeight cross-browser. This fixes #131
Additionally we unified the _reflow function and use the Math.max of the
document height and the max-keyframe in order to fix #133.