Add forceHeight: 'content' option for mobile
I kept having issues on mobile browsers in situations where I had
keyframes that didn't 'end' until after the end of the content. When
using forceHeight, the height is expanded to allow all keyframes to be
reachable by extending the page content beyond where it would normally
end. This small change adds a 'content' option for forceHeight which
measures the elements within the skrollr-body and adjusts the maximum
scroll area to be within the boundaries of the page content, regardless
of the final keyframe(s).